[api] Add threads and timestamps to api tests (#9349)
This commit builds threads and run ID on top of the consistent API testing introduced here. See proposal for more high level information.

Changes:

Refactored tests into dedicated file tests.rs.
Switched to creating accounts for every tests instead of using the same and created pre test set up routines in testsetups.rs.
Added run ID start_time to metrics.
