+++ crates/aptos-api-tester/src/counters.rs
@@ -7,49 +7,55 @@ pub static API_TEST_SUCCESS: Lazy<HistogramVec> = Lazy::new(|| {
     register_histogram_vec!(
         "api_test_success",
         "Number of user flows which succesfully passed",
-        &["test_name", "network_name"],
+        &["test_name", "network_name", "start_time"],
     )
     .unwrap()
 });
 
-pub fn test_success(test_name: &str, network_name: &str) -> Histogram {
-    API_TEST_SUCCESS.with_label_values(&[test_name, network_name])
+pub fn test_success(test_name: &str, network_name: &str, start_time: &str) -> Histogram {
+    API_TEST_SUCCESS.with_label_values(&[test_name, network_name, start_time])
 }
 
 pub static API_TEST_FAIL: Lazy<HistogramVec> = Lazy::new(|| {
     register_histogram_vec!(
         "api_test_fail",
         "Number of user flows which failed checks",
-        &["test_name", "network_name"],
+        &["test_name", "network_name", "start_time"],
     )
     .unwrap()
 });
 
-pub fn test_fail(test_name: &str, network_name: &str) -> Histogram {
-    API_TEST_FAIL.with_label_values(&[test_name, network_name])
+pub fn test_fail(test_name: &str, network_name: &str, start_time: &str) -> Histogram {
+    API_TEST_FAIL.with_label_values(&[test_name, network_name, start_time])
 }
 
 pub static API_TEST_ERROR: Lazy<HistogramVec> = Lazy::new(|| {
     register_histogram_vec!("api_test_error", "Number of user flows which crashed", &[
         "test_name",
         "network_name",
-    ])
+        "start_time"
+    ],)
     .unwrap()
 });
 
-pub fn test_error(test_name: &str, network_name: &str) -> Histogram {
-    API_TEST_ERROR.with_label_values(&[test_name, network_name])
+pub fn test_error(test_name: &str, network_name: &str, start_time: &str) -> Histogram {
+    API_TEST_ERROR.with_label_values(&[test_name, network_name, start_time])
 }
 
 pub static API_TEST_LATENCY: Lazy<HistogramVec> = Lazy::new(|| {
     register_histogram_vec!(
         "api_test_latency",
         "Time it takes to complete a user flow",
-        &["test_name", "network_name", "result"],
+        &["test_name", "network_name", "start_time", "result"],
     )
     .unwrap()
 });
 
-pub fn test_latency(test_name: &str, network_name: &str, result: &str) -> Histogram {
-    API_TEST_LATENCY.with_label_values(&[test_name, network_name, result])
+pub fn test_latency(
+    test_name: &str,
+    network_name: &str,
+    start_time: &str,
+    result: &str,
+) -> Histogram {
+    API_TEST_LATENCY.with_label_values(&[test_name, network_name, start_time, result])
 }
